refactor(api): PR-6 commit 5 — generate router 8 routes + daemon 控制
server.py 3192 → 2952 行（−240），8 generate routes + 1 BaseModel +
`_resolve_anima_model_paths` 共享 helper 抽出。

  api/routers/generate.py    8 routes  enqueue / get / taeflux status / install /
                                       daemon status / logs / unload / sample
  api/schemas/generate.py    1 BaseModel  GenerateRequest

  api/deps.py + _resolve_anima_model_paths()  reg_generate_prior 也用，
                                              升级到 deps 作共享

server.py：
  - `from .api.deps import _resolve_anima_model_paths, _supervisor` 添加
  - reg_generate_prior（仍在 server.py，等 PR-6.5 projects 一起搬）继续用同
    helper，零行为变更

回归：PR-1 安全网 152 + studio_server 44 + presets/import smoke = **229/229 全绿**。
route snapshot 不变（160 routes）。
