+++ studio/api/app.py
@@ -20,6 +20,7 @@ from .routers import (
     browse,
     data_exports,
     events_sse,
+    generate,
     health,
     installs,
     jobs,
@@ -59,3 +60,5 @@ app.include_router(upscalers.router)
 app.include_router(installs.router)
 # PR-6 commit 4: system router（11 routes: restart / update / rollback / preflight / etc.）
 app.include_router(system.router)
+# PR-6 commit 5: generate router（8 routes: 出图 + daemon 状态 + TAEFlux）
+app.include_router(generate.router)
