+++ studio/api/deps.py
@@ -25,3 +25,19 @@ def _supervisor() -> Supervisor:
     if sup is None:
         raise HTTPException(503, "supervisor not running")
     return sup
+
+
+def _resolve_anima_model_paths() -> dict[str, str]:
+    """解析 base 模型默认路径（先验生成 / 测试出图共用）。
+
+    与 version_config 的 model 字段对齐。用户用别的 base 模型时，
+    在 Settings → 模型 里改 selected_anima 影响这里的 anima 主权重路径。
+    """
+    from ..services.model_downloader import models_root
+    root = models_root()
+    return {
+        "transformer_path": str(root / "diffusion_models" / "anima-base-v1.0.safetensors"),
+        "vae_path": str(root / "vae" / "qwen_image_vae.safetensors"),
+        "text_encoder_path": str(root / "text_encoders"),
+        "t5_tokenizer_path": str(root / "t5_tokenizer"),
+    }
